Use u64 instead of u32 in DataStream for sizes and offsets

+++ Rush/UtilDataStream.h
@@ -14,32 +14,32 @@ class DataStream
 public:
 	virtual ~DataStream() {}
 
-	virtual u32               write(const void* buf, u32 size) = 0;
-	template <typename T> u32 writeT(const T& val) { return write(&val, sizeof(val)); }
+	virtual u64               write(const void* buf, u64 size) = 0;
+	template <typename T> u64 writeT(const T& val) { return write(&val, sizeof(val)); }
 
-	virtual u32 read(void* buf, u32 size) = 0;
+	virtual u64 read(void* buf, u64 size) = 0;
 
-	template <typename T> u32 readT(T& val) { return read(&val, sizeof(val)); }
+	template <typename T> u64 readT(T& val) { return read(&val, sizeof(val)); }
 
-	virtual u32  tell()             = 0;
-	virtual void seek(u32 pos)      = 0;
+	virtual u64  tell()             = 0;
+	virtual void seek(u64 pos)      = 0;
 	virtual void skip(int distance) = 0;
 	virtual void rewind()           = 0;
 
 	virtual bool valid() = 0;
 
-	virtual u32 length() = 0;
+	virtual u64 length() = 0;
 };
 
 class MemDataStream : public DataStream
 {
 public:
-	MemDataStream(void* data, u32 size) : m_dataRW((u8*)data), m_dataRO((const u8*)data), m_pos(0), m_size(size){};
-	MemDataStream(const void* data, u32 size) : m_dataRW(nullptr), m_dataRO((const u8*)data), m_pos(0), m_size(size){};
+	MemDataStream(void* data, u64 size) : m_dataRW((u8*)data), m_dataRO((const u8*)data), m_pos(0), m_size(size){};
+	MemDataStream(const void* data, u64 size) : m_dataRW(nullptr), m_dataRO((const u8*)data), m_pos(0), m_size(size){};
 
-	virtual u32 write(const void* buf, u32 size)
+	virtual u64 write(const void* buf, u64 size)
 	{
-		u32 pos2 = m_pos + size;
+		u64 pos2 = m_pos + size;
 		pos2     = min(pos2, m_size);
 		size     = pos2 - m_pos;
 
@@ -48,9 +48,9 @@ public:
 
 		return size;
 	}
-	virtual u32 read(void* buf, u32 size)
+	virtual u64 read(void* buf, u64 size)
 	{
-		u32 pos2 = m_pos + size;
+		u64 pos2 = m_pos + size;
 		pos2     = min(pos2, m_size);
 		size     = pos2 - m_pos;
 
@@ -60,11 +60,11 @@ public:
 		return size;
 	}
 
-	virtual u32  tell() { return m_pos; };
-	virtual void seek(u32 pos) { m_pos = min(pos, m_size); };
+	virtual u64  tell() { return m_pos; };
+	virtual void seek(u64 pos) { m_pos = min(pos, m_size); };
 	virtual void skip(int distance)
 	{
-		u32 pos2 = m_pos + distance;
+		u64 pos2 = m_pos + distance;
 		pos2     = min(pos2, m_size);
 		m_pos    = pos2;
 	};
@@ -72,14 +72,14 @@ public:
 
 	virtual bool valid() { return m_dataRO != nullptr; }
 
-	virtual u32 length() { return m_size; }
+	virtual u64 length() { return m_size; }
 
 private:
 	u8*       m_dataRW;
 	const u8* m_dataRO;
 
-	u32 m_pos;
-	u32 m_size;
+	u64 m_pos;
+	u64 m_size;
 };
 
 class NullDataStream : public DataStream
@@ -87,28 +87,28 @@ class NullDataStream : public DataStream
 public:
 	NullDataStream() : m_pos(0), m_size(0) {}
 
-	virtual u32 write(const void*, u32 size)
+	virtual u64 write(const void*, u64 size)
 	{
 		m_pos += size;
 		return m_pos;
 	}
-	virtual u32 read(void*, u32 size)
+	virtual u64 read(void*, u64 size)
 	{
 		m_pos += size;
 		return m_pos;
 	}
 
-	virtual u32  tell() { return m_pos; }
-	virtual void seek(u32 pos) { m_pos = pos; }
+	virtual u64  tell() { return m_pos; }
+	virtual void seek(u64 pos) { m_pos = pos; }
 	virtual void skip(int distance) { m_pos += distance; }
 	virtual void rewind() { m_pos = 0; }
 
 	virtual bool valid() { return true; }
 
-	virtual u32 length() { return m_size; }
+	virtual u64 length() { return m_size; }
 
 private:
-	u32 m_pos;
-	u32 m_size;
+	u64 m_pos;
+	u64 m_size;
 };
 }
